Project structure correction
